feat: initial tinkering
