+++ packages/web-extension/src/cip30/initializeInjectedScript.ts
@@ -1,6 +1,6 @@
 import { Cip30Wallet, WalletProperties, injectGlobal } from '@cardano-sdk/dapp-connector';
 import { Logger } from 'ts-log';
-import { MessengerDependencies, injectedRuntime } from '../messaging';
+import { MessengerDependencies, createInjectedRuntime } from '../messaging';
 import { consumeRemoteAuthenticatorApi } from './consumeRemoteAuthenticatorApi';
 import { consumeRemoteWalletApi } from './consumeRemoteWalletApi';
 
@@ -9,15 +9,21 @@ export interface InitializeInjectedDependencies {
 }
 
 // tested in e2e tests
-export const initializeInjectedScript = (props: WalletProperties, { logger }: InitializeInjectedDependencies) => {
-  const dependencies: MessengerDependencies = {
-    logger,
-    runtime: injectedRuntime
-  };
+// TODO: Should I create two different initialize Injected script functions? One for single API and another for multi-api
+export const initializeInjectedScript = (
+  props: Record<string, WalletProperties>,
+  { logger }: InitializeInjectedDependencies
+) => {
+  for (const [channelName, walletProps] of Object.entries(props)) {
+    const dependencies: MessengerDependencies = {
+      logger,
+      runtime: createInjectedRuntime(channelName)
+    };
 
-  const authenticator = consumeRemoteAuthenticatorApi(props, dependencies);
-  const walletApi = consumeRemoteWalletApi(props, dependencies);
-  const wallet = new Cip30Wallet(props, { api: walletApi, authenticator, logger });
+    const authenticator = consumeRemoteAuthenticatorApi(walletProps, dependencies);
+    const walletApi = consumeRemoteWalletApi(walletProps, dependencies);
+    const wallet = new Cip30Wallet(walletProps, { api: walletApi, authenticator, logger });
 
-  injectGlobal(window, wallet, logger);
+    injectGlobal(window, wallet, logger);
+  }
 };
